Tp Terminado

+++ src/main/java/lab1/meetNGame/WebRoutes.java
@@ -24,11 +24,6 @@ public class WebRoutes {
     public static final String LOGIN_TEMPLATE = "login.ftl";
     public static final String ADMIN_HOME_TEMPLATE = "adminhome.ftl";
     public static final String CREATE_GAME_TEMPLATE = "creategame.ftl";
-    public static final String UPDATE_GAME_TEMPLATE = "updategame.ftl";
-    public static final String DELETE_GAME_TEMPLATE = "deletegame.ftl";
-    public static final String MANAGE_RANK_TEMPLATE = "manageranks.ftl";
-    public static final String CREATE_RANK_TEMPLATE = "createrank.ftl";
-    public static final String DELETE_RANK_TEMPLATE = "deleterank.ftl";
     public static final String HOME_TEMPLATE = "home.ftl";
     public static final String ACCOUNT_SETTINGS_TEMPLATE = "settings.ftl";
     public static final String PROFILE_TEMPLATE = "profile.ftl";
@@ -41,7 +36,6 @@ public class WebRoutes {
     /**
      * ROUTES
      **/
-    //public static final String STATUS_ROUTE = "/status";
     public static final String REGISTER_ROUTE = "/register";
     public static final String LOGIN_ROUTE = "/login";
     public static final String LOGOUT_ROUTE = "/logout";
@@ -49,7 +43,6 @@ public class WebRoutes {
     public static final String CREATE_GAME_ROUTE = "/creategame";
     public static final String UPDATE_GAME_ROUTE = "/updategame";
     public static final String DELETE_GAME_ROUTE = "/deletegame";
-    public static final String MANAGE_RANK_ROUTE = "/manageranks";
     public static final String CREATE_RANK_ROUTE = "/createrank";
     public static final String DELETE_RANK_ROUTE = "/deleterank";
     public static final String HOME_ROUTE = "/home";
@@ -178,8 +171,6 @@ public class WebRoutes {
                     system.setMessage(null);
                 }
 
-                if (req.queryParams("deletedGame") != null) model.put("message", "Game deleted");
-                if (req.queryParams("updatedRanks") != null) model.put("message", "Ranks updated");
                 if (req.queryParams("deletedRanks") != null) model.put("message", "Rank deleted");
 
 
@@ -279,36 +270,21 @@ public class WebRoutes {
             }
         });
 
-        authenticatedGet(MANAGE_RANK_ROUTE, (req, res) -> {
-            final Optional<GamerUser> authenticatedGamerUser = getAuthenticatedGamerUser(req);
-            if(authenticatedGamerUser.get().isAdmin()){     //Se fija que no sea un gamer
-                return render(MANAGE_RANK_TEMPLATE);
-            }
-            else{                                           //Lleva a la pagina del gamer
-                final Map<String, Object> model = new HashMap<>();
-                String name = authenticatedGamerUser.get().getUserName();
-                model.put("message", "User is not Admin");
-                model.put("myName", name);
-                return render(model, HOME_TEMPLATE);
-            }
-        });
-
-        getGames(CREATE_RANK_ROUTE, CREATE_RANK_TEMPLATE);
-
         authenticatedPost(CREATE_RANK_ROUTE, (req, res) -> {
             final Optional<GamerUser> authenticatedGamerUser = getAuthenticatedGamerUser(req);
-            if(authenticatedGamerUser.get().isAdmin()){     //Se fija que no sea un gamer
+            if(authenticatedGamerUser.get().isAdmin()){                                         //Se fija que no sea un gamer
                 RankForm rankForm = RankForm.createFromBody(req.body());
                 Rank newRank = system.registerRank(rankForm.getGameName(), rankForm.getNewRank());
-                if (newRank != null){                       //Se fija que el rango no exista. Deberia fijarse que el juego existe
-                    res.redirect("/admin?updatedRanks");
+                if (newRank != null){                                                           //Se fija que el rango no exista. Deberia fijarse que el juego existe
+                    system.setMessage("Rank created");
+                    res.redirect("/admin?ok");
                 }
-                else {                                      //Vuelve al createrank con mensaje
+                else {                                                                          //Vuelve al createrank con mensaje
                     res.redirect("/createrank?rankexists");
                 }
                 return halt();
             }
-            else{                                           //Lleva a la pagina del gamer
+            else{                                                                               //Lleva a la pagina del gamer
                 final Map<String, Object> model = new HashMap<>();
                 String name = authenticatedGamerUser.get().getUserName();
                 model.put("message", "User is not Admin");
@@ -317,22 +293,20 @@ public class WebRoutes {
             }
         });
 
-        getGames(DELETE_RANK_ROUTE, DELETE_RANK_TEMPLATE);
-
         authenticatedPost(DELETE_RANK_ROUTE, (req, res) -> {
             final Optional<GamerUser> authenticatedGamerUser = getAuthenticatedGamerUser(req);
-            if(authenticatedGamerUser.get().isAdmin()){     //Se fija que no sea un gamer
+            if(authenticatedGamerUser.get().isAdmin()){                                         //Se fija que no sea un gamer
                 RankForm rankForm = RankForm.createFromBody(req.body());
                 Rank newRank = system.deleteRank(rankForm.getGameName(), rankForm.getNewRank());
-                if (newRank != null){                       //Se fija que el rango exista
+                if (newRank != null){                                                           //Se fija que el rango exista
                     res.redirect("/admin?deletedRanks");
                 }
-                else {                                      //Vuelve al deleterank con mensaje
+                else {                                                                          //Vuelve al deleterank con mensaje
                     res.redirect("/deleterank?rankNotExists");
                 }
                 return halt();
             }
-            else{                                           //Lleva a la pagina del gamer
+            else{                                                                              //Lleva a la pagina del gamer
                 final Map<String, Object> model = new HashMap<>();
                 String name = authenticatedGamerUser.get().getUserName();
                 model.put("message", "User is not Admin");
@@ -394,16 +368,11 @@ public class WebRoutes {
                 GamerDescription description = system.registerGamerDescription(authenticatedGamerUser.get(), myDescriptions, descriptionForm);
                 if (description != null){
                     res.redirect("/profile?ok");
-                    return halt();
                 }
                 else {
-                    List<GamerDescription> gamerDescriptions = system.getUserDescriptions(authenticatedGamerUser.get());
-                    List<Game> leftGames = system.getLeftGames(getGamesInDescription(gamerDescriptions));
-                    final String message = system.getMessage();
-                    final Map<String, Object> model = Map.of("message", message, "games", leftGames, "descriptions", gamerDescriptions);
-                    system.setMessage(null);
-                    return render(model, PROFILE_TEMPLATE);
+                    res.redirect("/profile?error");
                 }
+                return halt();
             }
             else {                                          //Lleva a la pagina del admin
                 final Map<String, Object> model = new HashMap<>();
@@ -417,39 +386,32 @@ public class WebRoutes {
             if (!authenticatedGamerUser.get().isAdmin()) {
                 UpdateDescriptionForm updateDescriptionForm = UpdateDescriptionForm.createFromBody(req.body());
                 if (updateDescriptionForm.getLvl().equals("") && updateDescriptionForm.getRank().equals("")) {
-                    List<GamerDescription> gamerDescriptions = system.getUserDescriptions(authenticatedGamerUser.get());
-                    List<Game> leftGames = system.getLeftGames(getGamesInDescription(gamerDescriptions));
-                    final Map<String, Object> model = Map.of("message", "You have to complete new lvl or rank", "games", leftGames, "descriptions", gamerDescriptions);
-                    return render(model, PROFILE_TEMPLATE);
+                    system.setMessage("Complete level or rank");
+                    res.redirect("/profile?error");
+                    return halt();
                 }
                 else if (!updateDescriptionForm.getLvl().equals("") && updateDescriptionForm.getRank().equals("")){
                     if (system.checkNewLevelDescription(updateDescriptionForm.getGameName(), updateDescriptionForm.getLvl())) {
                         system.updateDescriptionLvl(authenticatedGamerUser.get(), updateDescriptionForm.getGameName(), updateDescriptionForm.getLvl());
                         res.redirect("/profile?ok");
-                        return halt();
                     }
                     else {
-                        List<GamerDescription> gamerDescriptions = system.getUserDescriptions(authenticatedGamerUser.get());
-                        List<Game> leftGames = system.getLeftGames(getGamesInDescription(gamerDescriptions));
-                        final Map<String, Object> model = Map.of("message", system.getMessage(), "descriptions", gamerDescriptions, "games", leftGames);
-                        system.setMessage(null);
-                        return render(model, PROFILE_TEMPLATE);
+                        system.setMessage("Rank doesnt exist or does not belong to game");
+                        res.redirect("/profile?error");
                     }
+                    return halt();
                 }
                 else if (updateDescriptionForm.getLvl().equals("") && !updateDescriptionForm.getRank().equals("")){
                     Rank newRank = system.getRank(updateDescriptionForm.getRank(), updateDescriptionForm.getGameName());
                     if (newRank.getRankId() != null){
                         system.updateDescriptionRank(authenticatedGamerUser.get(), updateDescriptionForm.getGameName(),newRank);
                         res.redirect("/profile?ok");
-                        return halt();
                     }
                     else {
-                        List<GamerDescription> gamerDescriptions = system.getUserDescriptions(authenticatedGamerUser.get());
-                        List<Game> leftGames = system.getLeftGames(getGamesInDescription(gamerDescriptions));
-                        final Map<String, Object> model = Map.of("message","Rank doesnt exist or does not belong to game", "descriptions", gamerDescriptions, "games", leftGames);
-                        system.setMessage(null);
-                        return render(model, PROFILE_TEMPLATE);
+                        system.setMessage("Rank doesnt exist or does not belong to game");
+                        res.redirect("/profile?error");
                     }
+                    return halt();
                 }
                 else {
                     Rank newRank = system.getRank(updateDescriptionForm.getRank(), updateDescriptionForm.getGameName());
@@ -457,15 +419,12 @@ public class WebRoutes {
                         system.updateDescriptionLvl(authenticatedGamerUser.get(), updateDescriptionForm.getGameName(), updateDescriptionForm.getLvl());
                         system.updateDescriptionRank(authenticatedGamerUser.get(), updateDescriptionForm.getGameName(),newRank);
                         res.redirect("/profile?ok");
-                        return halt();
                     }
                     else {
-                        List<GamerDescription> gamerDescriptions = system.getUserDescriptions(authenticatedGamerUser.get());
-                        List<Game> leftGames = system.getLeftGames(getGamesInDescription(gamerDescriptions));
-                        final Map<String, Object> model = Map.of("message","Rank doesnt exist/Lvl is higher than lvlMax", "descriptions", gamerDescriptions, "games", leftGames);
-                        system.setMessage(null);
-                        return render(model, PROFILE_TEMPLATE);
+                        system.setMessage("Rank doesnt exist/Lvl is higher than lvlMax");
+                        res.redirect("/profile?error");
                     }
+                    return halt();
                 }
             } else {
                 final Map<String, Object> model = new HashMap<>();
@@ -476,10 +435,17 @@ public class WebRoutes {
 
         authenticatedPost(DELETE_DESCRIPTION_ROUTE, (req, res) -> {
             final Optional<GamerUser> authenticatedGamerUser = getAuthenticatedGamerUser(req);
-            SingleStringForm description = SingleStringForm.createFromBody(req.body());
-            system.deleteDescription(description.getLikedUser(), authenticatedGamerUser.get());
-            res.redirect("/profile?ok");
-            return halt();
+            if (!authenticatedGamerUser.get().isAdmin()) {
+                SingleStringForm description = SingleStringForm.createFromBody(req.body());
+                system.deleteDescription(description.getLikedUser(), authenticatedGamerUser.get());
+                res.redirect("/profile?ok");
+                return halt();
+            }
+            else {
+                final Map<String, Object> model = new HashMap<>();
+                model.put("message", "User is Admin");
+                return render(model, ADMIN_HOME_TEMPLATE);
+            }
         });
 
         authenticatedGet(MANAGE_INTEREST_ROUTE, (req, res) -> {
@@ -523,12 +489,8 @@ public class WebRoutes {
                     return halt();
                 }
                 else {
-                    List<GamerInterest> gamerInterests = system.getGamerInterest(authenticatedGamerUser.get());
-                    List<Game> leftGames = system.getLeftGames(getGamesInInterest(gamerInterests));
-                    final String message = system.getMessage();
-                    final Map<String, Object> model = Map.of("message", message, "games", leftGames, "interests", gamerInterests);
-                    system.setMessage(null);
-                    return render(model, MANAGE_INTEREST_TEMPLATE);
+                    res.redirect("/interests?error");
+                    return halt();
                 }
             }
             else {
@@ -610,10 +572,17 @@ public class WebRoutes {
 
         authenticatedPost(DELETE_INTEREST_ROUTE, (req, res) -> {
             final Optional<GamerUser> authenticatedGamerUser = getAuthenticatedGamerUser(req);
-            SingleStringForm description = SingleStringForm.createFromBody(req.body()); //uso LikeForm pq es igual
-            system.deleteInterest(description.getLikedUser(), authenticatedGamerUser.get());
-            res.redirect("/home?ok");
-            return halt();
+            if (!authenticatedGamerUser.get().isAdmin()) {
+                SingleStringForm description = SingleStringForm.createFromBody(req.body()); //uso LikeForm pq es igual
+                system.deleteInterest(description.getLikedUser(), authenticatedGamerUser.get());
+                res.redirect("/interests?ok");
+                return halt();
+            }
+            else {
+                final Map<String, Object> model = new HashMap<>();
+                model.put("message", "User is Admin");
+                return render(model, ADMIN_HOME_TEMPLATE);
+            }
         });
 
         authenticatedGet(FIND_PLAYERS_ROUTE, (req, res) -> {
@@ -729,6 +698,7 @@ public class WebRoutes {
             final Optional<GamerUser> currentUser = getAuthenticatedGamerUser(req);
             if (!currentUser.get().isAdmin()){
                 String username2 = req.params(":username");
+                String image2 = system.getUserImage(system.findUserByUserName(username2).get());
                 List<Message> messages = system.getMessages(currentUser.get().getUserName(), username2);
                 final Map<String, Object> model = new HashMap<>();
                 setImageAndNotif(currentUser, model);
@@ -737,6 +707,7 @@ public class WebRoutes {
                 }
                 model.put("sender", currentUser.get().getUserName());
                 model.put("receiver", username2);
+                model.put("image2", image2);
                 return new FreeMarkerEngine().render(new ModelAndView(model, "chat.ftl"));
             }
             else {
@@ -835,34 +806,6 @@ public class WebRoutes {
         List<Notification> notifications = system.getUserNotifications(currentUser.get().getUserName());
         model.put("notifications", notifications);
         model.put("image", image);
-    }
-
-    private void getGames(String gameRoute, String gameTemplate) {
-        authenticatedGet(gameRoute, (req, res) -> {
-            final Optional<GamerUser> authenticatedGamerUser = getAuthenticatedGamerUser(req);
-            if (authenticatedGamerUser.get().isAdmin()) {       //Se fija que no sea un gamer
-                List<Game> games = system.getGames();
-                final Map<String, Object> model = new HashMap<>();
-                if(!games.isEmpty()) {                          //Se fija que haya juegos creados
-                    model.put("games", games);
-                    if (req.queryParams("notOk") != null) model.put("message", "Fill at least one attribute to update");
-                    if (req.queryParams("noGameSelected") != null) model.put("message", "Choose a game to delete");
-                    if (req.queryParams("notGame") != null) model.put("message", "The game you chose does not exist");
-                    if (req.queryParams("rankexists") != null) model.put("message", "Rank name already exist");
-                    if (req.queryParams("rankNotExists") != null) model.put("message", "Rank name does not exist");
-                    return new FreeMarkerEngine().render(new ModelAndView(model, gameTemplate));
-                }else{                                          //Mensaje y lleva al createGame
-                    model.put("message", "There are no games to update");
-                    return render(model, CREATE_GAME_TEMPLATE);
-                }
-            } else {                                            //Lleva a la pagina del gamer
-                final Map<String, Object> model = new HashMap<>();
-                String name = authenticatedGamerUser.get().getUserName();
-                model.put("message", "User is not Admin");
-                model.put("myName", name);
-                return render(model, HOME_TEMPLATE);
-            }
-        });
     }
 
     private void authenticatedGet(String route, Route o) {
